Reverted all ID's
I tried to adjust the ID's so that the exported files from OTE were the same name as the files in the repository, but it just wouldn't work. The "verse" numbers got pushed down one level too much and all verses got the number "1". The correct one was still in the id, but OTE wasn't using that. It only works with three levels: book, chapter, verse.